+++ lib/src/get_command_p2.dart
@@ -24,10 +24,7 @@ class GetCommandP2<P1, P2> extends _GetCommandBase {
     super.dispose();
   }
 
-  /// A Command is a callable class.
-  ///
-  /// You can use Command() instead of Command.commandFunc() to execute the
-  /// command function.
+  /// Executes the command function [commandFunc].
   FutureOr<void> exec(P1 p1, P2 p2) {
     assert(commandFunc != null);
     return _doExecute(() => commandFunc!(p1, p2));